+++ backend/src/akvo/lumen/lib/pivot.clj
@@ -1,0 +1,7 @@
+(ns akvo.lumen.lib.pivot
+  (:require [akvo.lumen.lib.pivot-impl :as impl]))
+
+(defn query
+  "Return the result of a dataset pivot query"
+  [tenant-conn id query]
+  (impl/query tenant-conn id query))