+++ backend/src/akvo/lumen/lib/public.sql
@@ -1,4 +1,4 @@
 -- :name public-by-id :? :1
-SELECT *
+SELECT id, visualisation_id AS "visualisation-id", dashboard_id AS "dashboard-id"
 FROM share
 WHERE id = :id;