+++ client/src/components/dataset/sidebars/CombineColumns.jsx
@@ -87,10 +87,6 @@ export default class CombineColumns extends Component {
     return (
       <div
         className="DataTableSidebar"
-        style={{
-          width: '300px',
-          height: 'calc(100vh - 4rem)',
-        }}
       >
         <SidebarHeader onClose={onClose}>
           Combine columns