+++ client/src/containers/Visualisation/pivotTableSpecTemplate.js
@@ -1,0 +1,11 @@
+const pivotTableSpec = {
+  version: 1,
+  filters: [],
+  aggregation: 'count',
+  valueColumn: null,
+  categoryColumn: null,
+  rowColumn: null,
+  decimalPlaces: 3,
+};
+
+export default pivotTableSpec;